refactor: Moví todos los tipos de descuentos al módulo services.py
