+++ apps/comics/domain/services.py
@@ -1,6 +1,3 @@
-import uuid
-
-from . import agregates
 from .entities import Comic
 from .vo import ComicStatus
 
@@ -8,12 +5,62 @@ from .vo import ComicStatus
 class ObtainAmountToPayService:
     def __call__(self, comic: Comic):
         DISCOUNTS = {
-            ComicStatus.EXCELLENT: agregates.ExcellentComicDiscount.for_this(comic=comic).apply(),
-            ComicStatus.GOOD: agregates.GoodComicDiscount.for_this(comic=comic).apply(),
-            ComicStatus.ACCEPTABLE: agregates.AcceptableComicDiscount.for_this(comic=comic).apply(),
-            ComicStatus.IMPAIRED: agregates.ImpairedComicDiscount.for_this(comic=comic).apply(),
-            ComicStatus.DAMAGED: agregates.DamagedComicDiscount.for_this(comic=comic).apply()
+            ComicStatus.EXCELLENT: ExcellentComicDiscount.for_this(comic=comic).apply(),
+            ComicStatus.GOOD: GoodComicDiscount.for_this(comic=comic).apply(),
+            ComicStatus.ACCEPTABLE: AcceptableComicDiscount.for_this(comic=comic).apply(),
+            ComicStatus.IMPAIRED: ImpairedComicDiscount.for_this(comic=comic).apply(),
+            ComicStatus.DAMAGED: DamagedComicDiscount.for_this(comic=comic).apply()
         }
 
         discount = DISCOUNTS.get(comic.status)
         return comic.price - discount
+
+
+class Discount:
+    def __init__(self, comic: Comic):
+        self.comic = comic
+        self.percentage_of_discount = 0
+
+    @classmethod
+    def for_this(cls, comic: Comic):
+        return cls(comic=comic)
+
+    def apply(self):
+        return self.obtain_percentage(
+            price=self.comic.price,
+            percent=self.percentage_of_discount
+        )
+
+    @staticmethod
+    def obtain_percentage(price, percent):
+        return (price * percent) / 100
+
+
+class ExcellentComicDiscount(Discount):
+    def __init__(self, comic: Comic):
+        super().__init__(comic)
+        self.percentage_of_discount = 10
+
+
+class GoodComicDiscount(Discount):
+    def __init__(self, comic: Comic):
+        super().__init__(comic)
+        self.percentage_of_discount = 20
+
+
+class AcceptableComicDiscount(Discount):
+    def __init__(self, comic: Comic):
+        super().__init__(comic)
+        self.percentage_of_discount = 25
+
+
+class ImpairedComicDiscount(Discount):
+    def __init__(self, comic: Comic):
+        super().__init__(comic)
+        self.percentage_of_discount = 30
+
+
+class DamagedComicDiscount(Discount):
+    def __init__(self, comic: Comic):
+        super().__init__(comic)
+        self.percentage_of_discount = 50
